backend/requirements.txt
Django
django-environ
gunicorn
psycopg2-binary
django-rest-framework
djangorestframework
djangorestframework_simplejwt
django-cors-headers